fix(PostRepository) revert number of posts signature back to Long

+++ app/src/main/java/com/swent/skillswap/model/post/FakePostRepository.kt
@@ -52,7 +52,7 @@ class FakePostRepository : PostRepository {
     }
 
     override suspend fun getMultiplePosts(
-        numberOfPosts: Int,
+        numberOfPosts: Long,
         type: PostType,
         titleContains: String,
         ownerId: String,
